some fix

+++ internal/dto/dto.go
@@ -9,13 +9,13 @@ type AccrualResponse struct {
 	UserID      string  `json:"user_id"`
 	NumOrder    string  `json:"order"`
 	OrderStatus string  `json:"status"`
-	Accrual     float64 `json:"accrual"`
+	Accrual     float32 `json:"accrual"`
 }
 
 type Order struct {
 	NumOrder    string  `json:"number"`
 	OrderStatus string  `json:"status"`
-	Accrual     float64 `json:"accrual"`
+	Accrual     float32 `json:"accrual"`
 	UploadedAt  string  `json:"uploaded_at"`
 }
